refactor(cli): move CLI dispatch out of main.go into internal/cli
main.go now only wires os.Args / os.Stdout / os.Stderr into a single
entry point; argument parsing, subcommand dispatch, and usage rendering
live in internal/cli (cli.go, help.go, send.go, chat.go). The two
identical errSendUsage / errChatUsage sentinels collapse into a single
errUsage, and the top-level FlagSet switches from ExitOnError to
ContinueOnError so Run can return an exit code instead of calling
os.Exit mid-function. Behaviour is unchanged.
